+++ rewrite_emails.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-if [ "$GIT_AUTHOR_NAME" = "Moon9t" ]; then
-  export GIT_AUTHOR_EMAIL="luyanduthandokhumalo@gmail.com"
-  export GIT_COMMITTER_EMAIL="luyanduthandokhumalo@gmail.com"
-fi